Pass context to evaluate and call functions.
As of right now, there is no actual difference from the last
version, but this approach will eventually support callstacks
and make error handling easier to implement.

+++ include/lex.h
@@ -12,7 +12,7 @@
 // additional token values
 typedef struct token_info {
   char* string;
-  long double number;
+  double number;
 } token_info;
 
 typedef struct token {
